rename stared store into bookmark
