+++ ui/src/components/layout/SideBar.vue
@@ -119,7 +119,7 @@
 
     const menu = computed(() => {
         return [
-            ...(store.state.starred.pages?.length ? [{
+            ...(store.state.bookmarks.pages?.length ? [{
                 title: t("bookmark"),
                 icon: {
                     element: shallowRef(StarOutline),
@@ -127,7 +127,7 @@
                 },
                 child: [{
 
-                    component: () => h(BookmarkLinkList, {pages: store.state.starred.pages}),
+                    component: () => h(BookmarkLinkList, {pages: store.state.bookmarks.pages}),
                 }]
             }] : []),
             ...disabledCurrentRoute(props.generateMenu())
