+++ ui/src/stores/store.js
@@ -16,7 +16,7 @@ import taskrun from "./taskruns"
 import trigger from "./trigger";
 import editor from "./editor";
 import doc from "./doc";
-import starred from "./starred";
+import bookmarks from "./bookmarks";
 
 export default {
     modules: {
@@ -38,6 +38,6 @@ export default {
         trigger,
         editor,
         doc,
-        starred
+        bookmarks
     }
 }
